Eliminate shows.json intermediate file in build pipeline
generate-show-pages.py now fetches directly from the EIST API instead of
reading from data/shows.json. This removes an unnecessary intermediate step
and simplifies the build pipeline.

- Add API fetching with pagination to generate-show-pages.py
- Remove fetch-shows-from-api.py (now obsolete)
- Update build scripts to remove the fetch step
