+++ .github/workflows/deploy.yml
@@ -55,8 +55,6 @@ jobs:
         run: python3 generate-artist-pages.py
       - name: Align artist images for hero sections
         run: python3 scripts/detect-faces.py || true
-      - name: Fetch show data from API
-        run: python3 fetch-shows-from-api.py
       - name: Generate show pages
         run: python3 generate-show-pages.py
       - name: Cache Hugo files
